You can now get into the gameplay after authorization!!!
